refactor(tarball): split the crate root into per-concern modules
`pacquet-tarball`'s `lib.rs` had grown to 3251 lines holding six
unrelated concerns. Split along the seams that were already there:

- `error` — the error enums and the reqwest source-chain walker
- `download` — retry policy, integrity verification, progress events
- `extract` — gzip decode and tar entry extraction into the CAS
- `zip_archive` — the parallel path for runtime archives
- `prefetch` — the bulk store-index read
- `local_tarball` — reads of archives already on disk

The crate root keeps the shared statics, type aliases, and the
resolution-time fetch entry point, and re-exports each module so the
public API is unchanged.

Named `zip_archive` rather than `zip` so the module cannot shadow the
`zip` crate it calls into.

Pure code motion: no behavior change, verified by comparing every
content line of the original against the union of the new modules.
