+++ pnpm/crates/tarball/src/zip_archive.rs
@@ -1,0 +1,613 @@
+//! Zip-archive fetch and extraction.
+//!
+//! Runtime archives (Node.js / Bun / Deno) ship as zips rather than
+//! gzipped tarballs, so they take a parallel path to the tarball one:
+//! same CAS write and progress reporting, different container format.
+
+use super::{
+    Arc, AuthHeaders, CafsFileInfo, Component, Cursor, Duration, FetchingProgressLog,
+    FetchingProgressMessage, HashMap, HttpStatusError, IgnoreEntryFilter, Instant, Integrity,
+    LogEvent, LogLevel, MAX_UNTRUSTED_PREALLOC_BYTES, NetworkError, PackageFilesIndex, PathBuf,
+    PrefetchedCasPaths, ProgressLog, ProgressMessage, Read, Reporter, RequestRetryLog, RetryOpts,
+    SharedReadonlyStoreIndex, SharedVerifiedFilesCache, StoreDir, StoreIndexWriter, TarballError,
+    ThrottledClient, UNIX_EPOCH, VerifyChecksumError, allocate_tarball_buffer,
+    apply_append_manifest, apply_placeholder_manifest, emit_progress_found_in_store, file_mode,
+    is_transient_error, load_cached_cas_paths, post_download_semaphore, store_index_key,
+    tarball_error_to_request_retry,
+};
+
+/// Walk a zip archive, writing each regular-file entry into the CAFS
+/// and returning the `{relative-path → CAFS path}` map plus the
+/// per-package [`PackageFilesIndex`] row to hand off to the shared
+/// store-index writer. Matches the contract of [`crate::extract::extract_tarball_entries`]
+/// — same outputs, same per-file CAS write — but for binary
+/// `BinaryResolution { archive: zip, prefix: ... }` artifacts (the
+/// shape Node.js / Bun / Deno ships their Windows builds in).
+///
+/// 1. Directory entries are skipped; iterating only over file
+///    entries keeps the per-file ignore filter authoritative (a
+///    directory entry would otherwise expand to every descendant
+///    and bypass the filter).
+/// 2. Each entry's path is validated against absolute / `..`
+///    components via [`zip::read::ZipFile::enclosed_name`]. Any
+///    rejection is surfaced as [`TarballError::PathTraversal`] —
+///    carrying the `PATH_TRAVERSAL` error code.
+/// 3. If `archive_prefix` is set and the entry path starts with
+///    `{prefix}/`, the prefix is stripped before the ignore-filter
+///    check and before the entry is recorded in `cas_paths`, so the
+///    filter sees paths relative to the archive's top-level directory.
+/// 4. The cleaned path then runs through `ignore_file_pattern`;
+///    matching entries are dropped before any CAS write.
+/// 5. The remaining entry's bytes are read and committed via
+///    [`StoreDir::write_cas_file`], writing directly to the CAS.
+///
+/// Unix mode is read off the central-directory record via
+/// [`zip::read::ZipFile::unix_mode`]; archives written by Windows
+/// tooling don't populate it and we fall back to `0o644`.
+pub(crate) fn extract_zip_entries(
+    archive: &mut zip::ZipArchive<Cursor<Vec<u8>>>,
+    package_url: &str,
+    store_dir: &StoreDir,
+    archive_prefix: Option<&str>,
+    ignore_file_pattern: Option<&IgnoreEntryFilter>,
+) -> Result<(HashMap<String, PathBuf>, PackageFilesIndex), TarballError> {
+    let entry_count = archive.len();
+    let mut cas_paths = HashMap::<String, PathBuf>::with_capacity(entry_count);
+    let mut pkg_files_idx = PackageFilesIndex {
+        manifest: None,
+        requires_build: None,
+        algo: "sha512".to_string(),
+        files: HashMap::with_capacity(entry_count),
+        side_effects: None,
+    };
+
+    // Build the `{prefix}/` slice once. Treat `Some("")` as `None`,
+    // keeping entry paths verbatim when there is no prefix. The
+    // trailing slash anchors the strip so a prefix of `foo` doesn't
+    // accidentally consume `foobar/...`.
+    let basename_prefix: Option<String> =
+        archive_prefix.filter(|prefix| !prefix.is_empty()).map(|prefix| format!("{prefix}/"));
+
+    for i in 0..entry_count {
+        let mut entry = archive.by_index(i).map_err(|source| TarballError::ReadZipArchive {
+            url: package_url.to_string(),
+            source,
+        })?;
+        // Validate the path *before* the `is_dir()` early-skip so an
+        // archive carrying a directory entry like `../evil/` still
+        // surfaces [`TarballError::PathTraversal`] rather than being
+        // silently dropped. Pacquet wouldn't write that directory
+        // either way (only file entries take the CAS write path
+        // below), but rejecting outright keeps the "no unsafe entry
+        // accepted" contract intact for tooling that inspects the
+        // error code.
+        let raw_name = entry.name().to_string();
+        // [`zip::read::ZipFile::enclosed_name`] returns `None` for
+        // absolute paths and any path with a `..` component — a
+        // single check covers both forms of path traversal. The
+        // returned `PathBuf` has every `.` segment collapsed and is
+        // what we use below to build the canonical `cas_paths` /
+        // `pkg_files_idx` keys.
+        let Some(enclosed) = entry.enclosed_name() else {
+            return Err(TarballError::PathTraversal {
+                url: package_url.to_string(),
+                entry_path: raw_name,
+                reason: "zip entry path is absolute or escapes the archive root",
+            });
+        };
+        if entry.is_dir() {
+            continue;
+        }
+
+        // Rebuild the path into a forward-slash string from the
+        // sanitized `enclosed_name()` components. Three reasons over
+        // using the raw `entry.name()`:
+        //
+        // 1. `.` segments are already collapsed by `enclosed_name`,
+        //    so `pkg/./foo.txt` and `pkg/foo.txt` produce the same
+        //    `cas_paths` key — no accidental duplicates from
+        //    publisher tooling quirks.
+        // 2. The ignore filter sees the same canonical strings the
+        //    map is keyed by, so the regex / hand-coded matchers
+        //    can't be tripped up by `.` segments either.
+        // 3. Zip entries are spec'd to use `/` separators; this also
+        //    rejects any `\` an in-the-wild Windows-built archive
+        //    might have smuggled in (those would be `Normal`
+        //    components on Unix but interpreted as separators on
+        //    Windows).
+        //
+        // `enclosed_name` only yields `Normal` components, so the
+        // path-component walk below covers every case.
+        let normalized: String = enclosed
+            .components()
+            .map(|component| match component {
+                Component::Normal(name) => name.to_string_lossy().into_owned(),
+                _ => unreachable!("enclosed_name returns only Normal components: {:?}", enclosed),
+            })
+            .collect::<Vec<_>>()
+            .join("/");
+
+        // Strip the archive's top-level basename (`prefix` on
+        // `pacquet_lockfile::BinaryResolution`) so the ignore filter
+        // sees paths relative to the archive root. If the entry path
+        // doesn't start with `{prefix}/` we use the normalized form
+        // (a no-op when the entry already lives at the archive root).
+        let cleaned = match basename_prefix.as_deref() {
+            Some(prefix) => normalized.strip_prefix(prefix).unwrap_or(&normalized).to_string(),
+            None => normalized,
+        };
+        if cleaned.is_empty() {
+            // Skip an entry whose name was exactly the prefix
+            // directory: no relative payload survives the strip.
+            continue;
+        }
+
+        if let Some(filter) = ignore_file_pattern
+            && filter(&cleaned)
+        {
+            continue;
+        }
+
+        let prealloc_hint = entry.size().min(MAX_UNTRUSTED_PREALLOC_BYTES as u64) as usize;
+        let mut buffer = Vec::new();
+        buffer.try_reserve(prealloc_hint).map_err(|err| TarballError::ReadZipEntries {
+            url: package_url.to_string(),
+            entry_path: cleaned.clone(),
+            source: std::io::Error::new(
+                std::io::ErrorKind::OutOfMemory,
+                format!("failed to reserve {prealloc_hint} bytes for zip entry: {err}"),
+            ),
+        })?;
+        entry.read_to_end(&mut buffer).map_err(|source| TarballError::ReadZipEntries {
+            url: package_url.to_string(),
+            entry_path: cleaned.clone(),
+            source,
+        })?;
+
+        // Central-directory record carries a Unix mode only when
+        // the archive was built by a Unix tool; Windows-built
+        // archives omit it. Fall back to `0o644` so the executable
+        // bit defaults to off. Mask off the high `st_mode` bits
+        // (e.g. `0o100000` for a regular file) so `CafsFileInfo.mode`
+        // stays permission-only, matching the convention
+        // `add_files_from_dir.rs` enforces for tar / on-disk imports.
+        let file_mode = entry.unix_mode().unwrap_or(0o644) & 0o777;
+        let file_is_executable = file_mode::is_executable(file_mode);
+
+        let (file_path, file_hash) = store_dir
+            .write_cas_file(&buffer, file_is_executable)
+            .map_err(TarballError::WriteCasFile)?;
+
+        let file_size = u64::try_from(buffer.len()).unwrap_or(u64::MAX);
+        let checked_at =
+            UNIX_EPOCH.elapsed().ok().and_then(|elapsed| u64::try_from(elapsed.as_millis()).ok());
+        let file_attrs = CafsFileInfo {
+            digest: format!("{file_hash:x}"),
+            mode: file_mode,
+            size: file_size,
+            checked_at,
+        };
+
+        if let Some(previous) = cas_paths.insert(cleaned.clone(), file_path) {
+            tracing::warn!(?previous, "Duplication detected. Old entry has been ejected");
+        }
+        if let Some(previous) = pkg_files_idx.files.insert(cleaned, file_attrs) {
+            tracing::warn!(?previous, "Duplication detected. Old entry has been ejected");
+        }
+    }
+
+    Ok((cas_paths, pkg_files_idx))
+}
+
+/// Run one full zip-archive fetch attempt: hit the network, drain the
+/// body into RAM, verify the integrity hash, then walk the zip and
+/// extract every file entry into the CAFS. Mirrors
+/// [`crate::download::fetch_and_extract_once`] one-for-one (same network permit
+/// shape, same post-download semaphore gate, same retry-friendly
+/// errors) — only the `spawn_blocking` body differs: integrity check
+/// then [`extract_zip_entries`] instead of the gzip + tar path.
+///
+/// Writes directly into the CAS via [`StoreDir::write_cas_file`]
+/// rather than extracting to a temp dir and importing each file.
+// 8 arguments — over the default clippy threshold, but each is
+// distinct (see the matching note on `fetch_and_extract_zip_with_retry`).
+#[expect(
+    clippy::too_many_arguments,
+    reason = "the parameters are independent install-scoped inputs; bundling them into a struct only moves the same fields into a wrapper"
+)]
+pub(crate) async fn fetch_and_extract_zip_once<Reporter: self::Reporter>(
+    http_client: &ThrottledClient,
+    package_url: &str,
+    package_integrity: &Integrity,
+    package_id: &str,
+    attempt: u32,
+    store_dir: &'static StoreDir,
+    auth_headers: &AuthHeaders,
+    archive_prefix: Option<&str>,
+    ignore_file_pattern: Option<Arc<IgnoreEntryFilter>>,
+) -> Result<(HashMap<String, PathBuf>, PackageFilesIndex), TarballError> {
+    let network_error =
+        |error| TarballError::FetchTarball(NetworkError { url: package_url.to_string(), error });
+
+    let client = http_client.acquire_for_url(package_url).await;
+
+    let mut request = client.get(package_url);
+    // Match the tarball download path: resolve the per-URL auth
+    // header and attach it. Runtime artifacts (Node.js, Bun, Deno)
+    // are typically downloaded from public hosts that don't require
+    // auth, but a self-hosted mirror behind a token-protected proxy
+    // would 401 without this. Keeps parity with pnpm's binary
+    // fetcher which goes through the same `fetchFromRegistry` /
+    // auth-header plumbing.
+    if let Some(value) = auth_headers.for_url_with_package(package_url, Some(package_id)) {
+        request = request.header("authorization", value);
+    }
+
+    let send_result = request.send().await;
+    let size = send_result.as_ref().ok().and_then(reqwest::Response::content_length);
+    Reporter::emit(&LogEvent::FetchingProgress(FetchingProgressLog {
+        level: LogLevel::Debug,
+        message: FetchingProgressMessage::Started {
+            attempt: attempt + 1,
+            package_id: package_id.to_owned(),
+            size,
+        },
+    }));
+    let response_head = send_result.map_err(network_error)?;
+
+    let status = response_head.status();
+    if !status.is_success() {
+        const DRAIN_CAP: u64 = 64 * 1024;
+        if response_head.content_length().is_some_and(|len| len <= DRAIN_CAP) {
+            let _ = response_head.bytes().await;
+        }
+        return Err(TarballError::HttpStatus(HttpStatusError {
+            url: package_url.to_string(),
+            status: status.as_u16(),
+        }));
+    }
+
+    let expected_size = response_head.content_length();
+
+    let buffer = {
+        use futures_util::StreamExt;
+        let mut buf = allocate_tarball_buffer(expected_size, package_url)?;
+        let mut stream = response_head.bytes_stream();
+
+        const BIG_TARBALL_SIZE: u64 = 5 * 1024 * 1024;
+        const IN_PROGRESS_THROTTLE: Duration = Duration::from_millis(500);
+        let emit_progress = expected_size.is_some_and(|size| size >= BIG_TARBALL_SIZE);
+        let mut last_emit: Option<Instant> = None;
+        let mut last_emitted_downloaded: u64 = 0;
+        let mut downloaded: u64 = 0;
+        while let Some(chunk) = stream.next().await {
+            let chunk = chunk.map_err(network_error)?;
+            buf.extend_from_slice(&chunk);
+            downloaded = downloaded.saturating_add(chunk.len() as u64);
+            let throttle_ready =
+                last_emit.is_none_or(|instant| instant.elapsed() >= IN_PROGRESS_THROTTLE);
+            if emit_progress && throttle_ready {
+                Reporter::emit(&LogEvent::FetchingProgress(FetchingProgressLog {
+                    level: LogLevel::Debug,
+                    message: FetchingProgressMessage::InProgress {
+                        downloaded,
+                        package_id: package_id.to_owned(),
+                    },
+                }));
+                last_emit = Some(Instant::now());
+                last_emitted_downloaded = downloaded;
+            }
+        }
+        if emit_progress && downloaded != last_emitted_downloaded {
+            Reporter::emit(&LogEvent::FetchingProgress(FetchingProgressLog {
+                level: LogLevel::Debug,
+                message: FetchingProgressMessage::InProgress {
+                    downloaded,
+                    package_id: package_id.to_owned(),
+                },
+            }));
+        }
+        buf
+    };
+    drop(client);
+
+    let _post_download_permit = post_download_semaphore()
+        .acquire()
+        .await
+        .expect("post-download semaphore shouldn't be closed this soon");
+
+    tracing::info!(target: "pacquet::download", ?package_url, "Download completed");
+
+    let package_integrity = package_integrity.clone();
+    let package_url_owned = package_url.to_string();
+    let archive_prefix_owned: Option<String> = archive_prefix.map(str::to_string);
+    let result = tokio::task::spawn_blocking(
+        move || -> Result<(HashMap<String, PathBuf>, PackageFilesIndex), TarballError> {
+            package_integrity.check(&buffer).map_err(|error| {
+                TarballError::Checksum(VerifyChecksumError {
+                    url: package_url_owned.clone(),
+                    error,
+                })
+            })?;
+
+            // Open the archive in a scope so the buffer + ZipArchive
+            // are released before we return — large runtime archives
+            // (Node.js for Windows is ~30 MB) keep the buffer alive
+            // through the whole read otherwise.
+            let (cas_paths, pkg_files_idx) = {
+                let cursor = Cursor::new(buffer);
+                let mut archive = zip::ZipArchive::new(cursor).map_err(|source| {
+                    TarballError::ReadZipArchive { url: package_url_owned.clone(), source }
+                })?;
+                extract_zip_entries(
+                    &mut archive,
+                    &package_url_owned,
+                    store_dir,
+                    archive_prefix_owned.as_deref(),
+                    ignore_file_pattern.as_deref(),
+                )?
+            };
+            Ok((cas_paths, pkg_files_idx))
+        },
+    )
+    .await
+    .map_err(TarballError::TaskJoin)??;
+
+    tracing::info!(target: "pacquet::download", ?package_url, "Checksum verified");
+
+    Ok(result)
+}
+
+/// Run [`fetch_and_extract_zip_once`] under pnpm's retry policy.
+/// Same shape as [`crate::download::fetch_and_extract_with_retry`]: HTTP 401 / 403 /
+/// 404 fail fast, every other error retries with exponential
+/// backoff until [`RetryOpts::retries`] is exhausted. On success
+/// emits `pnpm:progress fetched` once per (resolved) package, same
+/// as the tarball path.
+// 10 arguments — over the default clippy threshold for the same
+// reason `fetch_and_extract_with_retry` is: each is distinct, and
+// bundling into a struct would just push the same fields into a
+// wrapper.
+#[expect(clippy::too_many_arguments, reason = "arg count is fixed by the fetcher signature")]
+pub(crate) async fn fetch_and_extract_zip_with_retry<Reporter: self::Reporter>(
+    http_client: &ThrottledClient,
+    package_url: &str,
+    package_integrity: &Integrity,
+    package_id: &str,
+    requester: &str,
+    store_dir: &'static StoreDir,
+    retry_opts: RetryOpts,
+    auth_headers: &AuthHeaders,
+    archive_prefix: Option<&str>,
+    ignore_file_pattern: Option<Arc<IgnoreEntryFilter>>,
+) -> Result<(HashMap<String, PathBuf>, PackageFilesIndex), TarballError> {
+    let mut attempt: u32 = 0;
+    loop {
+        let result = fetch_and_extract_zip_once::<Reporter>(
+            http_client,
+            package_url,
+            package_integrity,
+            package_id,
+            attempt,
+            store_dir,
+            auth_headers,
+            archive_prefix,
+            ignore_file_pattern.clone(),
+        )
+        .await;
+        match result {
+            Ok(value) => {
+                Reporter::emit(&LogEvent::Progress(ProgressLog {
+                    level: LogLevel::Debug,
+                    message: ProgressMessage::Fetched {
+                        package_id: package_id.to_owned(),
+                        requester: requester.to_owned(),
+                    },
+                }));
+                return Ok(value);
+            }
+            Err(err) if !is_transient_error(&err) => return Err(err),
+            Err(err) if attempt >= retry_opts.retries => {
+                tracing::warn!(
+                    target: "pacquet::download",
+                    ?package_url,
+                    attempts = attempt + 1,
+                    ?err,
+                    "Zip archive fetch retry budget exhausted",
+                );
+                return Err(err);
+            }
+            Err(err) => {
+                let delay = retry_opts.delay_for(attempt);
+                tracing::warn!(
+                    target: "pacquet::download",
+                    ?package_url,
+                    attempt = attempt + 1,
+                    max_attempts = retry_opts.retries + 1,
+                    ?delay,
+                    ?err,
+                    "Zip archive fetch failed; retrying after backoff",
+                );
+                Reporter::emit(&LogEvent::RequestRetry(RequestRetryLog {
+                    level: LogLevel::Debug,
+                    attempt: attempt + 1,
+                    error: tarball_error_to_request_retry(&err),
+                    max_retries: retry_opts.retries,
+                    method: "GET".to_string(),
+                    timeout: u64::try_from(delay.as_millis()).unwrap_or(u64::MAX),
+                    url: package_url.to_string(),
+                }));
+                tokio::time::sleep(delay).await;
+                attempt += 1;
+            }
+        }
+    }
+}
+
+/// Counterpart to [`crate::download::DownloadTarballToStore`] for zip-archive binary
+/// resolutions: the zip flow downloads the body, verifies the
+/// integrity hash, then walks zip entries and writes each to the CAFS
+/// — with the `prefix` field stripped from each entry path before the
+/// ignore filter and CAS write so the runtime's top-level
+/// `node-vX.Y.Z-<platform>-<arch>/` directory doesn't leak into
+/// downstream consumers' paths.
+///
+/// The store-index lookup, prefetch cache reuse, and store-index
+/// writer queueing match [`crate::download::DownloadTarballToStore`] — runtime
+/// artifacts share the same `index.db` schema as ordinary npm
+/// packages.
+#[must_use]
+pub struct DownloadZipArchiveToStore<'a> {
+    pub http_client: &'a ThrottledClient,
+    pub store_dir: &'static StoreDir,
+    pub store_index: Option<SharedReadonlyStoreIndex>,
+    pub store_index_writer: Option<Arc<StoreIndexWriter>>,
+    pub verify_store_integrity: bool,
+    pub verified_files_cache: SharedVerifiedFilesCache,
+    pub package_integrity: &'a Integrity,
+    pub package_url: &'a str,
+    pub package_id: &'a str,
+    pub requester: &'a str,
+    pub prefetched_cas_paths: Option<&'a PrefetchedCasPaths>,
+    pub retry_opts: RetryOpts,
+    /// Auth headers resolved at install start. The zip pipeline
+    /// applies the per-URL match the same way the tarball pipeline
+    /// does (`AuthHeaders::for_url`), so a runtime archive hosted
+    /// behind a token-protected proxy still authenticates correctly.
+    pub auth_headers: &'a AuthHeaders,
+    /// Basename of the archive's top-level directory, mirroring the
+    /// `prefix` field on `pacquet_lockfile::BinaryResolution`. The
+    /// zip extractor strips `{prefix}/` from each entry path before
+    /// the ignore-filter check and the CAS write, so downstream
+    /// consumers see paths relative to the package root rather than
+    /// the runtime-version-stamped wrapper directory.
+    pub archive_prefix: Option<&'a str>,
+    /// See [`crate::download::DownloadTarballToStore::ignore_file_pattern`] — the
+    /// per-fetch archive filter is shared by both archive types.
+    pub ignore_file_pattern: Option<Arc<IgnoreEntryFilter>>,
+    /// See [`crate::download::DownloadTarballToStore::offline`]. Same semantics for
+    /// the zip-archive path: when both cache lookups miss and
+    /// `offline` is `true`, the fetcher fails with
+    /// [`TarballError::NoOfflineTarball`] rather than hitting the
+    /// network.
+    pub offline: bool,
+    /// Synthesized `package.json` to fold into the extracted archive.
+    /// See [`crate::download::DownloadTarballToStore::append_manifest`] — same
+    /// `appendManifest` semantics for the zip path, used for the
+    /// runtime archives (e.g. Deno / Bun) that arrive as zips.
+    pub append_manifest: Option<&'a [u8]>,
+}
+
+impl DownloadZipArchiveToStore<'_> {
+    /// Execute the subroutine without an in-memory cache. Mirrors
+    /// [`crate::download::DownloadTarballToStore::run_without_mem_cache`] — same
+    /// prefetch-cas-paths reuse, same SQLite-index lookup, same
+    /// store-index writer queue — only the network and extract
+    /// path differs (zip instead of gzip + tar).
+    pub async fn run_without_mem_cache<Reporter: self::Reporter>(
+        &self,
+    ) -> Result<HashMap<String, PathBuf>, TarballError> {
+        let &DownloadZipArchiveToStore {
+            http_client,
+            store_dir,
+            package_integrity,
+            package_url,
+            package_id,
+            requester,
+            verify_store_integrity,
+            prefetched_cas_paths,
+            retry_opts,
+            auth_headers,
+            archive_prefix,
+            append_manifest,
+            ..
+        } = self;
+        let store_index = self.store_index.clone();
+        let store_index_writer = self.store_index_writer.clone();
+        let verified_files_cache = Arc::clone(&self.verified_files_cache);
+        // See the matching note in
+        // [`crate::download::DownloadTarballToStore::run_without_mem_cache`]: the
+        // Arc-wrapped filter can't ride along in the deref pattern,
+        // so clone it out by hand.
+        let ignore_file_pattern = self.ignore_file_pattern.clone();
+
+        let cache_key = store_index_key(&package_integrity.to_string(), package_id);
+        if let Some(prefetched) = prefetched_cas_paths
+            && let Some(cas_paths) = prefetched.get(&cache_key)
+        {
+            tracing::info!(
+                target: "pacquet::download",
+                ?package_url,
+                ?package_id,
+                "Reusing prefetched CAFS entry — skipping zip download",
+            );
+            emit_progress_found_in_store::<Reporter>(package_id, requester, None);
+            return Ok((**cas_paths).clone());
+        }
+        if let Some(cas_paths) = load_cached_cas_paths(
+            store_index,
+            store_dir,
+            cache_key,
+            verify_store_integrity,
+            verified_files_cache,
+        )
+        .await
+        {
+            tracing::info!(target: "pacquet::download", ?package_url, ?package_id, "Reusing cached CAFS entry — skipping zip download");
+            emit_progress_found_in_store::<Reporter>(package_id, requester, None);
+            return Ok(cas_paths);
+        }
+
+        // Offline-mode gate (zip archive). Same shape as the tarball
+        // path above — see the matching comment there for the
+        // rationale.
+        if self.offline {
+            tracing::warn!(
+                target: "pacquet::download",
+                ?package_url,
+                ?package_id,
+                "offline mode: zip archive missing from local store; refusing network fetch",
+            );
+            return Err(TarballError::NoOfflineTarball {
+                package_id: package_id.to_string(),
+                url: package_url.to_string(),
+            });
+        }
+
+        tracing::info!(target: "pacquet::download", ?package_url, "New cache (zip)");
+
+        let (mut cas_paths, mut pkg_files_idx) = fetch_and_extract_zip_with_retry::<Reporter>(
+            http_client,
+            package_url,
+            package_integrity,
+            package_id,
+            requester,
+            store_dir,
+            retry_opts,
+            auth_headers,
+            archive_prefix,
+            ignore_file_pattern,
+        )
+        .await?;
+
+        // Fold the synthesized runtime `package.json` into the row before
+        // it is persisted, so warm reinstalls (which read the row) get it.
+        if let Some(manifest_bytes) = append_manifest {
+            apply_append_manifest(store_dir, manifest_bytes, &mut cas_paths, &mut pkg_files_idx)?;
+        }
+        apply_placeholder_manifest(store_dir, &mut cas_paths, &mut pkg_files_idx)?;
+
+        let index_key = store_index_key(&package_integrity.to_string(), package_id);
+        if let Some(writer) = store_index_writer {
+            writer.queue(index_key, pkg_files_idx);
+        } else {
+            tracing::warn!(
+                target: "pacquet::download",
+                ?index_key,
+                "no shared store-index writer; skipping index row for this zip archive",
+            );
+        }
+
+        Ok(cas_paths)
+    }
+}
